refactor(water-diagram): move shared logic for flow propagation
