+++ process-flow-diagram-component/src/components/Diagram/FlowUtils.ts
@@ -3,9 +3,8 @@ import { DefaultEdgeOptions, EdgeTypes, ReactFlowInstance, Node, Edge, Connectio
 import BezierDiagramEdge from "../Edges/BezierDiagramEdge";
 import StraightDiagramEdge from "../Edges/StraightDiagramEdge";
 import StepDiagramEdge from "../Edges/StepDiagramEdge";
-import { NodeFlowProperty } from "./diagramReducer";
 import SmoothStepDiagramEdge from "../Edges/SmoothStepDiagramEdge";
-import { CustomEdgeData, DiagramCalculatedData, getNewProcessComponent, getNewNode, WaterProcessComponentType, UserDiagramOptions, ProcessFlowPart, getNewNodeId, FlowDiagramData, NodeFlowData, MAX_FLOW_DECIMALS } from "process-flow-lib";
+import { CustomEdgeData, DiagramCalculatedData, getNewProcessComponent, getNewNode, WaterProcessComponentType, UserDiagramOptions, ProcessFlowPart, getNewNodeId, FlowDiagramData, NodeFlowData, MAX_FLOW_DECIMALS, getEdgeSourceAndTarget, NodeFlowProperty } from "process-flow-lib";
 
 export const getRandomCoordinates = (height: number, width: number): { x: number, y: number } => {
   const screenWidth = window.innerWidth;
@@ -31,21 +30,6 @@ export const updateAssessmentCreatedNodes = (reactFlowInstance: ReactFlowInstanc
   return staleNodes;
 }
 
-export const getNodeFlowTotals = (connectedEdges: Edge[], nodes: Node[], selectedNodeId: string) => {
-  let totalCalculatedSourceFlow = 0;
-  let totalCalculatedDischargeFlow = 0;
-  connectedEdges.map((edge: Edge<CustomEdgeData>) => {
-    const { source, target } = getEdgeSourceAndTarget(edge, nodes);
-    if (selectedNodeId === target.diagramNodeId) {
-      totalCalculatedSourceFlow += edge.data.flowValue;
-    } else if (selectedNodeId === source.diagramNodeId) {
-      totalCalculatedDischargeFlow += edge.data.flowValue;
-    }
-  });
-
-  return { totalCalculatedSourceFlow, totalCalculatedDischargeFlow };
-}
-
 export const setCalculatedNodeDataProperty = (calculatedData: DiagramCalculatedData, nodeId: string, flowProperty: NodeFlowProperty, value: number) => {
   if (calculatedData.nodes[nodeId]) {
     calculatedData.nodes[nodeId][flowProperty] = value;
@@ -68,23 +52,6 @@ const setNodeFallbackPosition = (reactFlowInstance: ReactFlowInstance, node: Nod
   node.position = position;
 }
 
-
-export const getEdgeSourceAndTarget = (edge: Edge, nodes: Node[]) => {
-  let target: ProcessFlowPart;
-  let source: ProcessFlowPart;
-
-  nodes.forEach((node: Node) => {
-    if (node.id === edge.source) {
-      source = node.data as ProcessFlowPart;
-    }
-    if (node.id === edge.target) {
-      target = node.data  as ProcessFlowPart;
-    }
-  });
-
-  return { source, target };
-
-}
 
 export const getHasSources = (connectedEdges: Edge[], nodes: Node[], selectedNode: Node) => {
     return connectedEdges.some((edge: Edge) => {
